add filter roomate

+++ resources/js/components/Auth/Register.vue
@@ -8,17 +8,28 @@
                 class="object-cover w-full h-full"
             />
         </div>
-        <!-- Right: Login Form -->
+        <!-- Right: Sign Up Form -->
         <div class="lg:p-36 md:p-52 sm:20 p-8 w-full lg:w-1/2">
-            <h1 class="text-2xl font-semibold mb-4">Login</h1>
-            <form @submit.prevent="handleLogin">
-                <!-- Username Input -->
+            <h1 class="text-2xl font-semibold mb-4">Sign Up</h1>
+            <form @submit.prevent="handleSignup">
+                <!-- Name Input -->
+                <div class="mb-4">
+                    <label class="block text-gray-600">Name</label>
+                    <input
+                        type="text"
+                        id="name"
+                        v-model="name"
+                        class="w-full border border-gray-300 rounded-md py-2 px-3 focus:outline-none focus:border-blue-500"
+                        autocomplete="off"
+                    />
+                </div>
+                <!-- Email or Phone Input -->
                 <div class="mb-4">
                     <label class="block text-gray-600">Email or Phone</label>
                     <input
                         type="text"
-                        id="username"
-                        v-model="email"
+                        id="email_or_phone"
+                        v-model="emailOrPhone"
                         class="w-full border border-gray-300 rounded-md py-2 px-3 focus:outline-none focus:border-blue-500"
                         autocomplete="off"
                     />
@@ -31,43 +42,53 @@
                     <input
                         type="password"
                         id="password"
-                        name="password"
                         v-model="password"
                         class="w-full border border-gray-300 rounded-md py-2 px-3 focus:outline-none focus:border-blue-500"
                         autocomplete="off"
                     />
                 </div>
-                <!-- Remember Me Checkbox -->
+                <!-- Confirm Password Input -->
+                <div class="mb-4">
+                    <label for="password_confirm" class="block text-gray-600"
+                        >Confirm Password</label
+                    >
+                    <input
+                        type="password"
+                        id="password_confirm"
+                        v-model="passwordConfirm"
+                        class="w-full border border-gray-300 rounded-md py-2 px-3 focus:outline-none focus:border-blue-500"
+                        autocomplete="off"
+                    />
+                </div>
+                <!-- Terms and Conditions Checkbox -->
                 <div class="mb-4 flex items-center">
                     <input
                         type="checkbox"
-                        id="remember"
-                        name="remember"
+                        id="terms"
+                        v-model="acceptTerms"
                         class="text-blue-500"
                     />
-                    <label for="remember" class="text-gray-600 ml-2"
-                        >Remember Me</label
+                    <label for="terms" class="text-gray-600 ml-2"
+                        >I accept the Terms and Conditions</label
                     >
                 </div>
-                <!-- Forgot Password Link -->
-                <div class="mb-6 text-blue-500">
-                    <a href="#" class="hover:underline">Forgot Password?</a>
-                </div>
-                <!-- Login Button -->
+                <!-- Sign Up Button -->
                 <button
                     type="submit"
                     class="bg-blue-500 hover:bg-blue-600 text-white font-semibold rounded-md py-2 px-4 w-full"
                 >
-                    Login
+                    Sign Up
                 </button>
             </form>
-            <!-- Sign up  Link -->
+            <!-- Sign up Link -->
             <div class="mt-6 text-blue-500 text-center">
-                <router-link to="register" class="hover:underline"
-                    >Sign up Here</router-link
-                >
+                <p>
+                    Already have an account?
+                    <router-link to="login" class="hover:underline"
+                        >Login Here</router-link
+                    >
+                </p>
             </div>
-
             <!-- Return Button -->
             <div class="mt-4 text-center">
                 <button
@@ -83,29 +104,36 @@
 
 <script>
 import { defaultApiClient } from "../../axios";
+
 export default {
     data() {
         return {
-            email: "",
+            name: "",
+            emailOrPhone: "",
             password: "",
+            passwordConfirm: "",
+            acceptTerms: false,
         };
     },
     methods: {
-        goHome() {
-            this.$router.push("/"); // Chuyển hướng về trang chủ
-        },
-        async handleLogin() {
-            console.log(this.email);
+        async handleSignup() {
+            // Kiểm tra nếu password và confirm password không trùng khớp
+            if (this.password !== this.passwordConfirm) {
+                alert("Passwords do not match!");
+                return;
+            }
 
             try {
-                // Gửi yêu cầu đăng nhập và đợi phản hồi từ API
-                const response = await defaultApiClient.post("/login", {
-                    email: this.email,
+                // Gửi yêu cầu đăng ký và đợi phản hồi từ API
+                const response = await defaultApiClient.post("/register", {
+                    name: this.name,
+                    email: this.emailOrPhone,
                     password: this.password,
+                    password_confirmation: this.passwordConfirm,
                 });
 
-                // Xử lý khi đăng nhập thành công
-                console.log("Đăng nhập thành công:", response.data.data);
+                // Xử lý khi đăng ký thành công
+                console.log("Đăng ký thành công:", response.data.data);
                 localStorage.setItem(
                     "user",
                     JSON.stringify(response.data.user)
@@ -115,19 +143,24 @@ export default {
                     response.data.data.access_token
                 );
 
-                // Ví dụ: chuyển hướng người dùng hoặc lưu token đăng nhập
+                // Chuyển hướng người dùng sau khi đăng ký thành công
                 if (response.data.roles.includes("Admin")) {
                     this.$router.push("/admin");
                 } else {
                     this.$router.push("/");
                 }
             } catch (error) {
-                // Xử lý lỗi đăng nhập
+                // Xử lý lỗi đăng ký
                 console.error(
-                    "Đăng nhập thất bại:",
+                    "Đăng ký thất bại:",
                     error.response?.data || error.message
                 );
             }
+        },
+
+        // Phương thức để quay về trang chủ
+        goHome() {
+            this.$router.push("/"); // Chuyển hướng về trang chủ
         },
     },
 };
